split work between tasks, implemented spi dma and buffer queues

+++ Inc/xensiv_bgt60trxx.h
@@ -254,7 +254,7 @@ int32_t xensiv_bgt60trxx_set_fifo_limit(const xensiv_bgt60trxx_t* dev,
  * an error indicating what went wrong.
  */
 int32_t xensiv_bgt60trxx_get_fifo_data(xensiv_bgt60trxx_t* dev,
-                                       uint16_t* data,
+                                       uint8_t* data,
                                        uint32_t num_samples);
 
 /**
